# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Competiscope v2 API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/runs": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "List Runs",
        "operationId": "list_runs_api_runs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunSummary"
                  },
                  "type": "array",
                  "title": "Response List Runs Api Runs Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Create Run",
        "operationId": "create_run_api_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/RunDetail"
                    },
                    {
                      "$ref": "#/components/schemas/WorkflowStartResponse"
                    }
                  ],
                  "title": "Response Create Run Api Runs Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Run",
        "operationId": "get_run_api_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/runs/{run_id}/trace": {
      "get": {
        "tags": [
          "trace"
        ],
        "summary": "Get Trace",
        "operationId": "get_trace_api_runs__run_id__trace_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RunEvent"
                  },
                  "title": "Response Get Trace Api Runs  Run Id  Trace Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/trace/spans": {
      "get": {
        "tags": [
          "trace"
        ],
        "summary": "Get Trace Spans",
        "operationId": "get_trace_spans_api_runs__run_id__trace_spans_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TraceSpan"
                  },
                  "title": "Response Get Trace Spans Api Runs  Run Id  Trace Spans Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/trace/otel": {
      "get": {
        "tags": [
          "trace"
        ],
        "summary": "Get Trace Otel Export",
        "operationId": "get_trace_otel_export_api_runs__run_id__trace_otel_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtelTraceExport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/trace/observability": {
      "get": {
        "tags": [
          "trace"
        ],
        "summary": "Get Trace Observability Report",
        "operationId": "get_trace_observability_report_api_runs__run_id__trace_observability_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TraceObservabilityReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/compliance": {
      "get": {
        "tags": [
          "trace"
        ],
        "summary": "Get Run Compliance Report",
        "operationId": "get_run_compliance_report_api_runs__run_id__compliance_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunComplianceReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/trace/agent-messages": {
      "get": {
        "tags": [
          "trace"
        ],
        "summary": "Get Agent Messages",
        "operationId": "get_agent_messages_api_runs__run_id__trace_agent_messages_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AgentMessage"
                  },
                  "title": "Response Get Agent Messages Api Runs  Run Id  Trace Agent Messages Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/trace/tool-calls": {
      "get": {
        "tags": [
          "trace"
        ],
        "summary": "Get Tool Call Messages",
        "operationId": "get_tool_call_messages_api_runs__run_id__trace_tool_calls_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ToolCallMessage"
                  },
                  "title": "Response Get Tool Call Messages Api Runs  Run Id  Trace Tool Calls Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RunCreateRequest": {
        "properties": {
          "workspace_id": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Workspace Id",
            "default": "default-workspace"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 160,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "topic": {
            "type": "string",
            "maxLength": 200,
            "minLength": 2,
            "title": "Topic"
          },
          "competitors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 8,
            "title": "Competitors"
          },
          "dimensions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 8,
            "minItems": 1,
            "title": "Dimensions"
          },
          "competitor_layer": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "L1",
                  "L2",
                  "L3"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Competitor Layer"
          },
          "scenario_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario Id"
          },
          "execution_mode": {
            "type": "string",
            "enum": [
              "auto",
              "demo",
              "real"
            ],
            "title": "Execution Mode",
            "default": "auto"
          },
          "auto_redo_warn_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto Redo Warn Enabled"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "topic",
          "dimensions"
        ],
        "title": "RunCreateRequest"
      },
      "RunDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "idempotency_key": {
            "type": "string",
            "title": "Idempotency Key",
            "default": ""
          },
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id",
            "default": "default-workspace"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "interrupted",
              "completed",
              "failed"
            ],
            "title": "Status"
          },
          "execution_mode": {
            "type": "string",
            "enum": [
              "demo",
              "real"
            ],
            "title": "Execution Mode"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "plan": {
            "$ref": "#/components/schemas/AnalysisPlan"
          },
          "max_iterations": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Max Iterations",
            "default": 2
          },
          "auto_redo_warn_enabled": {
            "type": "boolean",
            "title": "Auto Redo Warn Enabled",
            "default": false
          },
          "report_md": {
            "type": "string",
            "title": "Report Md",
            "default": ""
          },
          "raw_sources": {
            "items": {
              "$ref": "#/components/schemas/RawSource"
            },
            "type": "array",
            "title": "Raw Sources"
          },
          "competitor_kbs": {
            "additionalProperties": {
              "$ref": "#/components/schemas/CompetitorKB"
            },
            "type": "object",
            "title": "Competitor Kbs"
          },
          "competitor_knowledge": {
            "additionalProperties": {
              "$ref": "#/components/schemas/CompetitorKnowledge"
            },
            "type": "object",
            "title": "Competitor Knowledge"
          },
          "competitor_discovery": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CompetitorDiscovery"
              },
              {
                "type": "null"
              }
            ]
          },
          "comparison_matrix": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ComparisonMatrix"
              },
              {
                "type": "null"
              }
            ]
          },
          "qa_findings": {
            "items": {
              "$ref": "#/components/schemas/QCIssue"
            },
            "type": "array",
            "title": "Qa Findings"
          },
          "reflections": {
            "items": {
              "$ref": "#/components/schemas/ReflectionRecord"
            },
            "type": "array",
            "title": "Reflections"
          },
          "revisions": {
            "items": {
              "$ref": "#/components/schemas/RevisionRecord"
            },
            "type": "array",
            "title": "Revisions"
          },
          "agent_messages": {
            "items": {
              "$ref": "#/components/schemas/AgentMessage"
            },
            "type": "array",
            "title": "Agent Messages"
          },
          "tool_call_messages": {
            "items": {
              "$ref": "#/components/schemas/ToolCallMessage"
            },
            "type": "array",
            "title": "Tool Call Messages"
          },
          "trace_spans": {
            "items": {
              "$ref": "#/components/schemas/TraceSpan"
            },
            "type": "array",
            "title": "Trace Spans"
          },
          "metrics": {
            "$ref": "#/components/schemas/RunMetrics"
          },
          "current_node": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Node"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "topic",
          "status",
          "execution_mode",
          "created_at",
          "updated_at",
          "plan"
        ],
        "title": "RunDetail"
      },
      "RunEvent": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id",
            "default": ""
          },
          "type": {
            "type": "string",
            "enum": [
              "run_created",
              "node_started",
              "node_completed",
              "interrupt",
              "qa_issue",
              "report_updated",
              "revision_recorded",
              "run_completed",
              "run_failed"
            ],
            "title": "Type"
          },
          "agent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent"
          },
          "subagent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subagent"
          },
          "swimlane": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Swimlane"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "run_id",
          "type",
          "message"
        ],
        "title": "RunEvent"
      },
...
      "RunSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "idempotency_key": {
            "type": "string",
            "title": "Idempotency Key",
            "default": ""
          },
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id",
            "default": "default-workspace"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "interrupted",
              "completed",
              "failed"
            ],
            "title": "Status"
          },
          "execution_mode": {
            "type": "string",
            "enum": [
              "demo",
              "real"
            ],
            "title": "Execution Mode"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "topic",
          "status",
          "execution_mode",
          "created_at",
          "updated_at"
        ],
        "title": "RunSummary"
      },
...
      "TraceSpan": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id",
            "default": ""
          },
          "otel_span_id": {
            "type": "string",
            "title": "Otel Span Id",
            "default": ""
          },
          "parent_span_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Span Id"
          },
          "traceparent": {
            "type": "string",
            "title": "Traceparent",
            "default": ""
          },
          "kind": {
            "type": "string",
            "enum": [
              "llm",
              "search",
              "fetch",
              "tool"
            ],
            "title": "Kind"
          },
          "agent": {
            "type": "string",
            "title": "Agent"
          },
          "subagent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subagent"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "type": "string",
            "enum": [
              "ok",
              "error"
            ],
            "title": "Status"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "duration_ms": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Duration Ms"
          },
          "input_chars": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Input Chars",
            "default": 0
          },
          "output_chars": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Output Chars",
            "default": 0
          },
          "input_tokens_estimate": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Input Tokens Estimate",
            "default": 0
          },
          "output_tokens_estimate": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Output Tokens Estimate",
            "default": 0
          },
          "cost_estimate_usd": {
            "type": "number",
            "minimum": 0.0,
            "title": "Cost Estimate Usd",
            "default": 0.0
          },
          "input_preview": {
            "type": "string",
            "title": "Input Preview",
            "default": ""
          },
          "output_preview": {
            "type": "string",
            "title": "Output Preview",
            "default": ""
          },
          "full_input": {
            "type": "string",
            "title": "Full Input",
            "default": ""
          },
          "full_output": {
            "type": "string",
            "title": "Full Output",
            "default": ""
          },
          "metadata": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                },
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Metadata"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "kind",
          "agent",
          "name",
          "status",
          "duration_ms"
        ],
        "title": "TraceSpan"
      },
...
      "WorkflowStartResponse": {
        "properties": {
          "workflow_id": {
            "type": "string",
            "title": "Workflow Id"
          },
          "workflow_type": {
            "type": "string",
            "const": "CompetitiveIntelWorkflow",
            "title": "Workflow Type",
            "default": "CompetitiveIntelWorkflow"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "idempotency_key": {
            "type": "string",
            "title": "Idempotency Key"
          },
          "task_queue": {
            "type": "string",
            "title": "Task Queue"
          },
          "status": {
            "type": "string",
            "enum": [
              "started",
              "already_started"
            ],
            "title": "Status"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "workflow_id",
          "run_id",
          "idempotency_key",
          "task_queue",
          "status"
        ],
        "title": "WorkflowStartResponse"
      },
...
    }
  }
}